+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 VERSION 0.2
 -------------
 
+* Refactor MakeBaseImage.sh to BuildImage.sh
+
 * Webcam support
 
 * OpenGL support